+++ lib/src/extensions.dart
@@ -1,4 +1,5 @@
 import 'package:meta/meta.dart';
+import 'package:clock/clock.dart';
 
 extension NumTimeExtension<T extends num> on T {
   /// Returns a Duration represented in weeks
@@ -33,6 +34,115 @@ extension DateTimeTimeExtension on DateTime {
   /// Subtracts the Duration from this DateTime returns the difference as a new DateTime object.
   DateTime operator -(Duration duration) => subtract(duration);
 
+  /// Returns only year, month and day
+  DateTime get date => DateTime(year, month, day);
+
+  /// Returns only the time
+  Duration get timeOfDay => hour.hours + minute.minutes + second.seconds;
+
+  /// Returns if today, true
+  bool get isToday {
+    return _calculateDifference(this) == 0;
+  }
+
+  /// Returns if tomorrow, true
+  bool get isTomorrow {
+    return _calculateDifference(this) == 1;
+  }
+
+  /// Returns if yesterday, true
+  bool get wasYesterday {
+    return _calculateDifference(this) == -1;
+  }
+
+  /// Returns true if this year is a leap year.
+  bool get isLeapYear =>
+      // Leap years are used since 1582.
+      year >= 1582 && year % 4 == 0 && (year % 100 != 0 || year % 400 == 0);
+
+  /// Returns the amount of days that are in this month.
+  ///
+  /// Accounts for leap years.
+  int get daysInMonth {
+    final days = [
+      31, // January
+      if (isLeapYear) 29 else 28, // February
+      31, // March
+      30, // April
+      31, // May
+      30, // June
+      31, // July
+      31, // August
+      30, // September
+      31, // October
+      30, // November
+      31, // December
+    ];
+
+    return days[month - 1];
+  }
+
+  /// Returns true if [other] is in the same year as [this].
+  ///
+  /// Does not account for timezones.
+  bool isAtSameYearAs(DateTime other) => year == other.year;
+
+  /// Returns true if [other] is in the same month as [this].
+  ///
+  /// This means the exact month, including year.
+  ///
+  /// Does not account for timezones.
+  bool isAtSameMonthAs(DateTime other) => isAtSameYearAs(other) && month == other.month;
+
+  /// Returns true if [other] is on the same day as [this].
+  ///
+  /// This means the exact day, including year and month.
+  ///
+  /// Does not account for timezones.
+  bool isAtSameDayAs(DateTime other) => isAtSameMonthAs(other) && day == other.day;
+
+  /// Returns true if [other] is at the same hour as [this].
+  ///
+  /// This means the exact hour, including year, month and day.
+  ///
+  /// Does not account for timezones.
+  bool isAtSameHourAs(DateTime other) => isAtSameDayAs(other) && hour == other.hour;
+
+  /// Returns true if [other] is at the same minute as [this].
+  ///
+  /// This means the exact minute, including year, month, day and hour.
+  ///
+  /// Does not account for timezones.
+  bool isAtSameMinuteAs(DateTime other) => isAtSameHourAs(other) && minute == other.minute;
+
+  /// Returns true if [other] is at the same second as [this].
+  ///
+  /// This means the exact second, including year, month, day, hour and minute.
+  ///
+  /// Does not account for timezones.
+  bool isAtSameSecondAs(DateTime other) => isAtSameMinuteAs(other) && second == other.second;
+
+  /// Returns true if [other] is at the same millisecond as [this].
+  ///
+  /// This means the exact millisecond,
+  /// including year, month, day, hour, minute and second.
+  ///
+  /// Does not account for timezones.
+  bool isAtSameMillisecondAs(DateTime other) => isAtSameSecondAs(other) && millisecond == other.millisecond;
+
+  /// Returns true if [other] is at the same microsecond as [this].
+  ///
+  /// This means the exact microsecond,
+  /// including year, month, day, hour, minute, second and millisecond.
+  ///
+  /// Does not account for timezones.
+  bool isAtSameMicrosecondAs(DateTime other) => isAtSameMillisecondAs(other) && microsecond == other.microsecond;
+
+  static int _calculateDifference(DateTime date) {
+    final now = clock.now();
+    return DateTime(date.year, date.month, date.day).difference(DateTime(now.year, now.month, now.day)).inDays;
+  }
+
   /// Returns a range of dates to [to], exclusive start, inclusive end
   /// ```dart
   /// final start = DateTime(2019);
@@ -64,25 +174,91 @@ extension DateTimeTimeExtension on DateTime {
   }
 
   DateTime copyWith({
-    int year,
-    int month,
-    int day,
-    int hour,
-    int minute,
-    int second,
-    int millisecond,
-    int microsecond,
+    int? year,
+    int? month,
+    int? day,
+    int? hour,
+    int? minute,
+    int? second,
+    int? millisecond,
+    int? microsecond,
   }) {
-    return DateTime(
-      year ?? this.year,
-      month ?? this.month,
-      day ?? this.day,
-      hour ?? this.hour,
-      minute ?? this.minute,
-      second ?? this.second,
-      millisecond ?? this.millisecond,
-      microsecond ?? this.microsecond,
+    return isUtc
+        ? DateTime.utc(
+            year ?? this.year,
+            month ?? this.month,
+            day ?? this.day,
+            hour ?? this.hour,
+            minute ?? this.minute,
+            second ?? this.second,
+            millisecond ?? this.millisecond,
+            microsecond ?? this.microsecond,
+          )
+        : DateTime(
+            year ?? this.year,
+            month ?? this.month,
+            day ?? this.day,
+            hour ?? this.hour,
+            minute ?? this.minute,
+            second ?? this.second,
+            millisecond ?? this.millisecond,
+            microsecond ?? this.microsecond,
+          );
+  }
+
+  /// Returns the Monday of this week
+  DateTime get firstDayOfWeek =>
+      isUtc ? DateTime.utc(year, month, day + 1 - weekday) : DateTime(year, month, day + 1 - weekday);
+
+  /// Returns the Sunday of this week
+  DateTime get lastDayOfWeek =>
+      isUtc ? DateTime.utc(year, month, day + 7 - weekday) : DateTime(year, month, day + 7 - weekday);
+
+  /// Returns the first day of this month
+  DateTime get firstDayOfMonth => isUtc ? DateTime.utc(year, month, 1) : DateTime(year, month, 1);
+
+  /// Returns the last day of this month (considers leap years)
+  DateTime get lastDayOfMonth => isUtc ? DateTime.utc(year, month + 1, 0) : DateTime(year, month + 1, 0);
+
+  /// Returns the first day of this year
+  DateTime get firstDayOfYear => isUtc ? DateTime.utc(year, 1, 1) : DateTime(year, 1, 1);
+
+  /// Returns the last day of this year
+  DateTime get lastDayOfYear => isUtc ? DateTime.utc(year, 12, 31) : DateTime(year, 12, 31);
+
+  /// Returns this [DateTime] clamped to be in the range [min]-[max].
+  ///
+  /// The comparison is done using [compareTo].
+  ///
+  /// The arguments [min] and [max] must form a valid range where
+  /// `min.compareTo(max) <= 0`.
+  ///
+  /// Example:
+  /// ```dart
+  /// var result = DateTime(2022, DateTime.october, 15).clamp(
+  ///   min: DateTime(2022, DateTime.september, 1),
+  ///   max: DateTime(2022, DateTime.september, 30),
+  /// ); // DateTime(2022, DateTime.september, 30);
+  /// result = DateTime(2022, DateTime.august, 21).clamp(
+  ///   min: DateTime(2022, DateTime.september, 15),
+  ///   max: DateTime(2022, DateTime.september, 30),
+  /// ); // DateTime(2022, DateTime.september, 15);
+  /// result = DateTime(2022, DateTime.september, 1).clamp(
+  ///   min: DateTime(2022, DateTime.august, 1),
+  ///   max: DateTime(2022, DateTime.september, 30),
+  /// ); // DateTime(2022, DateTime.september, 1);
+  /// ```
+  DateTime clamp({DateTime? min, DateTime? max}) {
+    assert(
+      ((min != null) && (max != null)) ? min.compareTo(max).isNegative : true,
+      'DateTime min has to be before max\n(min: $min - max: $max)',
     );
+    if ((min != null) && compareTo(min).isNegative) {
+      return min;
+    } else if ((max != null) && max.compareTo(this).isNegative) {
+      return max;
+    }
+    return this;
   }
 
   /// Adds months
@@ -165,14 +341,46 @@ extension DurationTimeExtension on Duration {
   int get inWeeks => (inDays / daysPerWeek).ceil();
 
   /// Adds the Duration to the current DateTime and returns a DateTime in the future
-  DateTime get fromNow => DateTime.now() + this;
-
-  @Deprecated('Use fromNow instead. Will be removed in 2.0.0')
-  DateTime get later => fromNow;
+  DateTime get fromNow => clock.now() + this;
 
   /// Subtracts the Duration from the current DateTime and returns a DateTime in the past
-  DateTime get ago => DateTime.now() - this;
+  DateTime get ago => clock.now() - this;
 
   /// Returns a Future.delayed from this
   Future<void> get delay => Future.delayed(this);
+
+  /// Returns this [Duration] clamped to be in the range [min]-[max].
+  ///
+  /// The comparison is done using [compareTo].
+  ///
+  /// The arguments [min] and [max] must form a valid range where
+  /// `min.compareTo(max) <= 0`.
+  ///
+  /// Example:
+  /// ```dart
+  /// var result = Duration(days: 10, hours: 12).clamp(
+  ///   min: Duration(days: 5),
+  ///   max: Duration(days: 10),
+  /// ); // Duration(days: 10)
+  /// result = Duration(hours: 18).clamp(
+  ///   min: Duration(days: 5),
+  ///   max: Duration(days: 10),
+  /// ); // Duration(days: 5)
+  /// result = Duration(days: 0).clamp(
+  ///   min: Duration(days: -5),
+  ///   max: Duration(days: 5),
+  /// ); // Duration(days: 0)
+  /// ```
+  Duration clamp({Duration? min, Duration? max}) {
+    assert(
+      ((min != null) && (max != null)) ? min.compareTo(max).isNegative : true,
+      'Duration min has to be shorter than max\n(min: $min - max: $max)',
+    );
+    if ((min != null) && compareTo(min).isNegative) {
+      return min;
+    } else if ((max != null) && max.compareTo(this).isNegative) {
+      return max;
+    }
+    return this;
+  }
 }